requirements.txt
Django==5.0
djangorestframework==3.14.0
django-jazzmin==2.6.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
Pillow==10.1.0
mysqlclient==2.2.1
requests==2.32.3
astral==3.2
